Deleting unneeded files. Moving resource-list component to the main components tab.
