+++ app/volunteer/page.tsx
@@ -1,33 +1,24 @@
-import {
-    Table,
-    TableBody,
-    TableCaption,
-    TableCell,
-    TableHead,
-    TableHeader,
-    TableRow,
-} from "@/components/ui/table"
+import { createClient } from "@/utils/supabase/server";
+import ResourceList from "@/components/resource-list";
 
-export default function Page() {
-  return (
-    <Table>
-      <TableCaption>A list of your recent invoices.</TableCaption>
-      <TableHeader>
-        <TableRow>
-          <TableHead className="w-[100px]">Invoice</TableHead>
-          <TableHead>Status</TableHead>
-          <TableHead>Method</TableHead>
-          <TableHead className="text-right">Amount</TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        <TableRow>
-          <TableCell className="font-medium">INV001</TableCell>
-          <TableCell>Paid</TableCell>
-          <TableCell>Credit Card</TableCell>
-          <TableCell className="text-right">$250.00</TableCell>
-        </TableRow>
-      </TableBody>
-    </Table>
-  )
+export default async function Page() {
+  const supabase = await createClient();  // Await the client creation
+
+  // Get current timestamp
+  const now = new Date().toISOString();
+
+  // Fetch only active resources
+  const { data: resources, error } = await supabase
+    .from("fire_resources")
+    .select("*")
+    .gte("active_end", now)  // active_end ≥ now
+    .lte("active_start", now); // active_start ≤ now
+
+  if (error) {
+    console.error("Error fetching resources:", error);
+    return <p>Failed to load resources.</p>;
+  }
+
+  return <ResourceList initialResources={resources} />;
 }
+
