+++ components/resource-list.tsx
